+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log for Terraform AWS Kinesis Firehose Splunk
 
+## v8.1.0
+ * Making var.name_cloudwatch_logs_to_ship non-mandatory. The var will now default to `null` and the subscription filter will not be created in such case.
+ * Adding var.cloudwatch_logs_group_names_to_ship to allow creating subscription filters to multiple log groups.
+
 ## v8.0.0 - **Breaking Changes**
  * Requires `>= 5.0.0, < 6.0.0` of the terraform aws [provider](https://registry.terraform.io/providers/hashicorp/aws/latest/docs)
  * Updated default value of `var.nodejs_runtime` to `nodejs18.x`